+++ package.json
@@ -4,13 +4,16 @@
   "private": true,
   "scripts": {
     "build:client": "webpack",
+    "build:server": "babel src --out-dir dist",
+    "build": "npm run build:client && npm run build:server",
     "start:client": "webpack --watch",
-    "start:server": "nodemon --exec babel-node ./bin/www",
+    "start:server": "nodemon --exec babel-node src/server/app.js",
     "start": "npm run start:client & npm run start:server"
   },
   "dependencies": {
     "axios": "^0.23.0",
     "bcrypt": "^5.0.1",
+    "body-parser": "^1.19.0",
     "bootstrap": "^4.6.0",
     "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
