+++ src/server/helpers/generateToken.js
