+++ src/server/helpers/handlePassword.js
