+++ webpack.config.js
@@ -3,7 +3,7 @@ const path = require('path');
 module.exports = {
     mode: 'development',
 
-    context: path.resolve(__dirname,'client'),
+    context: path.resolve(__dirname,'src/client'),
 
     entry: {
         app: './index.js'
